backend/requirements.txt
Flask
Flask-Cors
python-dotenv
openai
transformers
torch
pydub
